core: FTSState moved to common_constants Fix #4241
